refactor: Move code from biUtils to budget-insight
Budget-insight is the interface BI cozy
bi-http is only the the http interface to BI
